Clean up the liquidation code
What was happening before was stupid. The liquidation penalty was being
applied to the collateral (you'd get 10% more collateral than you
bought). This applies the liquidation penalty to the price. The
liquidator now purchases $100 tokens for $90. This is consistent with
how Aave handles it and also more beneficial to the liquidators.