Creates a rinkeby deploy
Refactors the deploy steps a little bit. Merges the oracle steps into a
single step. Also will deploy a stub for the chainlink OHM/ETH
aggregator, so we can test the full oracle.

NOTE: I've generated a market, but I'm not sure how to get the market
contract to be verified.