ZIM-01M: Inexistent Validation of Collateral Extracted
Allow liquidators to set a minimum amount of collateral they recieve
when liquidating an account. If they don't wish to use this, they can
set it to 0. This is important because liquidating results in an oracle
update, so there may be a price change while they are waiting in the
mempool.